+++ api/src/main/java/io/opentelemetry/distributedcontext/EmptyDistributedContext.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.dctx;
+package io.opentelemetry.distributedcontext;
 
 import java.util.Collections;
 import java.util.Iterator;
