+++ api/src/main/java/io/opentelemetry/distributedcontext/unsafe/ContextUtils.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.dctx.unsafe;
+package io.opentelemetry.distributedcontext.unsafe;
 
 import io.grpc.Context;
 import io.opentelemetry.context.Scope;
-import io.opentelemetry.dctx.DistributedContext;
-import io.opentelemetry.dctx.EmptyDistributedContext;
+import io.opentelemetry.distributedcontext.DistributedContext;
+import io.opentelemetry.distributedcontext.EmptyDistributedContext;
 
 /**
  * Utility methods for accessing the {@link DistributedContext} contained in the {@link
