+++ api/src/test/java/io/opentelemetry/OpenTelemetryTest.java
@@ -22,10 +22,10 @@ import static org.junit.Assert.assertTrue;
 import io.opentelemetry.context.Scope;
 import io.opentelemetry.context.propagation.BinaryFormat;
 import io.opentelemetry.context.propagation.HttpTextFormat;
-import io.opentelemetry.dctx.DefaultDistributedContextManager;
-import io.opentelemetry.dctx.DistributedContext;
-import io.opentelemetry.dctx.DistributedContextManager;
-import io.opentelemetry.dctx.spi.DistributedContextManagerProvider;
+import io.opentelemetry.distributedcontext.DefaultDistributedContextManager;
+import io.opentelemetry.distributedcontext.DistributedContext;
+import io.opentelemetry.distributedcontext.DistributedContextManager;
+import io.opentelemetry.distributedcontext.spi.DistributedContextManagerProvider;
 import io.opentelemetry.metrics.CounterDouble;
 import io.opentelemetry.metrics.CounterLong;
 import io.opentelemetry.metrics.DefaultMeter;
