--- sdk/src/main/resources/META-INF/services/io.opentelemetry.dctx.spi.DistributedContextManagerProvider
@@ -1,1 +1,0 @@
-io.opentelemetry.sdk.dctx.DistributedContextManagerSdkProvider
