+++ sdk/src/main/resources/META-INF/services/io.opentelemetry.distributedcontext.spi.DistributedContextManagerProvider
@@ -1,0 +1,1 @@
+io.opentelemetry.sdk.distributedcontext.DistributedContextManagerSdkProvider
